+++ app/src/js/components/Granules/bulk-granule-delete.js
@@ -1,38 +1,66 @@
-import React from 'react';
+import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 
+import { bulkGranule } from '../../actions';
 import _config from '../../config';
 import DefaultModal from '../Modal/modal';
 import TextArea from '../TextAreaForm/text-area';
 
 const { kibanaRoot } = _config;
 
+const defaultQuery = {
+  index: '',
+  query: '',
+  forceRemoveFromCmr: false
+};
+
 const BulkDeleteModal = ({
-  showModal,
-  onChange,
-  onConfirm,
-  onCancel,
-  inflight,
-  query,
-  success,
   asyncOpId,
+  className,
+  dispatch,
   error,
-  selected
+  handleSuccessConfirm,
+  inflight,
+  onCancel,
+  requestId,
+  selected,
+  showModal,
+  success
 }) => {
+  const [query, setQuery] = useState(JSON.stringify(defaultQuery, null, 2));
+  const [errorState, setErrorState] = useState();
+
   const buttonText = inflight ? 'loading...'
     : success ? 'Success!' : 'Run Bulk Delete';
+  const formError = error || errorState;
+
+  function handleSubmit (e) {
+    e.preventDefault();
+    if (status !== 'inflight') {
+      try {
+        var json = JSON.parse(query);
+      } catch (e) {
+        return setErrorState('Syntax error in JSON');
+      }
+      dispatch(bulkGranule({ requestId, json }));
+    }
+  }
+
+  function onChange (id, value) {
+    setQuery(value);
+  }
 
   return (
     <DefaultModal
       title='Bulk Granule Delete'
-      className='bulk_granules'
+      className={className}
       showModal={showModal}
       cancelButtonText={success ? 'Close' : 'Cancel Bulk Delete'}
       confirmButtonText={success ? 'Go To Operations' : buttonText}
       confirmButtonClass='button__bulkgranules'
       onCancel={onCancel}
       onCloseModal={onCancel}
-      onConfirm={onConfirm}
+      onConfirm={success ? handleSuccessConfirm : handleSubmit}
     >
       {success &&
         <p>
@@ -45,6 +73,8 @@ const BulkDeleteModal = ({
           <h4 className="modal_subtitle">To run and complete your bulk delete task:</h4>
           <p>
             1. In the box below, add either an array of granule Ids or an elasticsearch query and index. <br/>
+            2. Set <strong>forceRemoveFromCmr</strong> to <strong>true</strong> to automatically have granules removed from CMR as part of deletion.<br/>
+            If <strong>forceRemoveFromCmr</strong> is <strong>false</strong>, then the bulk granule deletion will <strong>fail for any granules that are published to CMR.</strong>
           </p>
           {selected &&
             <>
@@ -64,7 +94,7 @@ const BulkDeleteModal = ({
             <TextArea
               value={query}
               id='run-bulk-granule'
-              error={error}
+              error={formError}
               onChange={onChange}
               mode='json'
               minLines={30}
@@ -78,17 +108,17 @@ const BulkDeleteModal = ({
 };
 
 BulkDeleteModal.propTypes = {
-  showModal: PropTypes.bool,
-  onChange: PropTypes.func,
-  onCloseModal: PropTypes.func,
-  onConfirm: PropTypes.func,
-  onCancel: PropTypes.func,
-  inflight: PropTypes.bool,
-  success: PropTypes.bool,
   asyncOpId: PropTypes.string,
+  className: PropTypes.string,
+  dispatch: PropTypes.func,
   error: PropTypes.string,
+  handleSuccessConfirm: PropTypes.func,
+  inflight: PropTypes.bool,
+  onCancel: PropTypes.func,
+  requestId: PropTypes.string,
   selected: PropTypes.array,
-  query: PropTypes.string
+  showModal: PropTypes.bool,
+  success: PropTypes.bool
 };
 
 export default BulkDeleteModal;
